[Exchange] simplify & clean ccxt exchange code and move logic to connector
tmp


tmp


tmp


tmp
