+++ octobot_trading/exchanges/__init__.py
@@ -117,6 +117,7 @@ from octobot_trading.exchanges.implementations import (
     MarginExchangeSimulator,
     MarginCCXTExchange,
     CryptofeedWebSocketExchange,
+    CCXTExchangeCommons,
 )
 from octobot_trading.exchanges import exchange_builder
 from octobot_trading.exchanges.exchange_builder import (
@@ -170,6 +171,7 @@ __all__ = [
     "CCXTWebSocketExchange",
     "DefaultCCXTSpotExchange",
     "CryptofeedWebSocketExchange",
+    "CCXTExchangeCommons",
     "SpotExchangeSimulator",
     "SpotCCXTExchange",
     "FutureExchangeSimulator",
