+++ octobot_trading/exchanges/implementations/__init__.pxd
@@ -46,6 +46,10 @@ from octobot_trading.exchanges.implementations cimport cryptofeed_websocket_exch
 from octobot_trading.exchanges.implementations.cryptofeed_websocket_exchange cimport (
     CryptofeedWebSocketExchange,
 )
+from octobot_trading.exchanges.implementations cimport ccxt_exchange_commons
+from octobot_trading.exchanges.implementations.ccxt_exchange_commons cimport (
+    CCXTExchangeCommons,
+)
 
 __all__ = [
     "SpotExchangeSimulator",
@@ -56,4 +60,5 @@ __all__ = [
     "SpotCCXTExchange",
     "CCXTWebSocketExchange",
     "CryptofeedWebSocketExchange",
+    "CCXTExchangeCommons",
 ]
